refactor: Cookie에 대한 Session을 관리하는 CookieSession 클래스 생성
