+++ tomcat/src/main/java/org/apache/coyote/http11/session/SessionParser.java
@@ -7,7 +7,7 @@ public class SessionParser {
 
     private static final String COOKIE_SESSION_KEY = "JSESSIONID";
 
-    public static Optional<Session> extractSessionFromRequest(final HttpRequest request) {
+    public static Optional<Session> extractCookieSessionFromRequest(final HttpRequest request) {
         final Optional<String> sessionId = request.getCookie().get(COOKIE_SESSION_KEY);
         if (sessionId.isEmpty()) {
             return Optional.empty();
